+++ packages/queue/src/jobsManagement/commands/workerHealthCommand.ts
@@ -1,0 +1,10 @@
+import {
+  WorkerStatusWithId,
+  WorkerRepository,
+} from "../../worker/workerRepository";
+
+export function workerHealthCommand(): WorkerStatusWithId[] {
+  const statuses = WorkerRepository.Instance.getWorkerStatuses();
+
+  return statuses;
+}