+++ packages/queue/src/reporter/totalCount/ITotalCountReporter.ts
@@ -1,0 +1,5 @@
+export interface ITotalCountReporter {
+  setTotal(total: number): void;
+  increment(n?: number): void;
+  reset(value?: number): void;
+}