+++ packages/queue/src/types.ts
@@ -7,9 +7,32 @@ export {
   WorkerOptions,
   Queue as RawQueue,
   QueueOptions,
+  JobState,
+  JobType,
   JobsOptions,
 } from "bullmq";
 
-// This may seems odd but actual Job data type from bullmq is any 🤷‍♂️
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-export type JobData = any;
+export type JobData = unknown;
+
+export type JobReturnValue = unknown;
+
+export enum JobStateEnum {
+  ACTIVE = "active",
+  DELAYED = "delayed",
+  WAITING = "waiting",
+  COMPLETED = "completed",
+  FAILED = "failed",
+  WAITING_CHILDREN = "waiting-children",
+}
+
+export enum JobTypeEnum {
+  ACTIVE = "active",
+  DELAYED = "delayed",
+  WAITING = "waiting",
+  COMPLETED = "completed",
+  FAILED = "failed",
+  WAITING_CHILDREN = "waiting-children",
+  PAUSED = "paused",
+  REPEAT = "repeat",
+  WAIT = "wait",
+}